tidy up router
