+++ packages/app/src/Router.tsx
@@ -1,16 +1,12 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { extractUrlValue } from '@w3ux/utils'
 import { PagesConfig } from 'config/pages'
-import { useActiveAccounts } from 'contexts/ActiveAccounts'
-import { useImportedAccounts } from 'contexts/Connect/ImportedAccounts'
-import { useOtherAccounts } from 'contexts/Connect/OtherAccounts'
 import { useNetwork } from 'contexts/Network'
 import { usePlugins } from 'contexts/Plugins'
 import { useStaking } from 'contexts/Staking'
 import { useUi } from 'contexts/UI'
-import { Notifications } from 'controllers/Notifications'
+import { useAccountFromUrl } from 'hooks/useAccountFromUrl'
 import { ErrorFallbackApp, ErrorFallbackRoutes } from 'library/ErrorBoundary'
 import { Headers } from 'library/Headers'
 import { Help } from 'library/Help'
@@ -25,7 +21,6 @@ import { Tooltip } from 'library/Tooltip'
 import { Overlays } from 'overlay'
 import { useEffect, useRef } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
-import { useTranslation } from 'react-i18next'
 import {
   HashRouter,
   Navigate,
@@ -37,19 +32,13 @@ import { StakingApi } from 'StakingApi'
 import { Body, Main } from 'ui-structure'
 
 const RouterInner = () => {
-  const { t } = useTranslation()
   const { network } = useNetwork()
   const { inSetup } = useStaking()
   const { pathname } = useLocation()
   const { setContainerRefs } = useUi()
   const { pluginEnabled } = usePlugins()
-  const { accounts } = useImportedAccounts()
-  const { accountsInitialised } = useOtherAccounts()
-  const { activeAccount, setActiveAccount } = useActiveAccounts()
-  const nominating = !inSetup()
-  const stakingApiEnabled = pluginEnabled('staking_api')
 
-  // References to outer container.
+  // References to outer container
   const mainInterfaceRef = useRef<HTMLDivElement>(null)
 
   // Scroll to top of the window on every page change or network change
@@ -64,29 +53,12 @@ const RouterInner = () => {
     })
   }, [])
 
-  // Open default account modal if url var present and accounts initialised
-  useEffect(() => {
-    if (accountsInitialised) {
-      const aUrl = extractUrlValue('a')
-      if (aUrl) {
-        const account = accounts.find((a) => a.address === aUrl)
-        if (account && aUrl !== activeAccount) {
-          setActiveAccount(account.address || null)
-
-          Notifications.emit({
-            title: t('accountConnected', { ns: 'library' }),
-            subtitle: `${t('connectedTo', { ns: 'library' })} ${
-              account.name || aUrl
-            }.`,
-          })
-        }
-      }
-    }
-  }, [accountsInitialised])
+  // Support active account from url
+  useAccountFromUrl()
 
   return (
     <ErrorBoundary FallbackComponent={ErrorFallbackApp}>
-      {stakingApiEnabled && nominating && <StakingApi />}
+      {pluginEnabled('staking_api') && !inSetup() && <StakingApi />}
       <NotificationPrompts />
       <Body>
         <Help />
